+++ packages/messaging/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mysten/messaging
 
+## 0.1.0
+
+### Minor Changes
+
+- bfeb536: Introduce the `addMembers` API to enable channel creators to add members to existing channels
+
+  Expose three new methods following the SDK pattern:
+  - addMembers(): Transaction builder
+  - addMembersTransaction(): Returns Transaction object
+  - executeAddMembersTransaction(): Execute and return results with member details
+
 ## 0.0.3
 
 ### Patch Changes